+++ CHANGELOG.md
@@ -19,8 +19,9 @@
 ## [WIP] 0.4.0-alpha2 / 2017-August-??
 
 - Merge context from `bract-context.dev.edn` in dev mode when available
-- Catch exception and print stack trace in dev mode (due to global uncaught handler)
+- Catch exception and print stack trace in dev mode (due to uncaught handler)
 - Key definition for context-inducers now defaults to an empty vector
+- Add utility function `pst-when-uncaught-handler` for printing stack trace
 
 
 ## 0.4.0-alpha1 / 2017-July-31